+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* fixed result with empty list
 * added optional type in prepare statment
 
 ## 1.5.6